+++ apps/aktivitetspenger-soknad/env.schema.ts
@@ -10,6 +10,8 @@ import * as z from 'zod';
 export enum AppEnvKey {
     'VELG_SCENARIO' = 'VELG_SCENARIO',
     'SIF_PUBLIC_URL_SKATTEETATEN' = 'SIF_PUBLIC_URL_SKATTEETATEN',
+    'SIF_PUBLIC_URL_PERSONOPPLYSNINGER' = 'SIF_PUBLIC_URL_PERSONOPPLYSNINGER',
+    'SIF_PUBLIC_URL_ENDRE_KONTONUMMER' = 'SIF_PUBLIC_URL_ENDRE_KONTONUMMER',
     'SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL' = 'SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL',
     'SIF_PUBLIC_USE_FARO' = 'SIF_PUBLIC_USE_FARO',
 }
@@ -19,6 +21,8 @@ export const appEnvSchema = z
         [AppEnvKey.VELG_SCENARIO]: z.string().optional(),
         [AppEnvKey.SIF_PUBLIC_USE_FARO]: z.string().optional(),
         [AppEnvKey.SIF_PUBLIC_NAIS_FRONTEND_TELEMETRY_COLLECTOR_URL]: z.string().optional(),
+        [AppEnvKey.SIF_PUBLIC_URL_PERSONOPPLYSNINGER]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_ENDRE_KONTONUMMER]: z.string(),
         [AppEnvKey.SIF_PUBLIC_URL_SKATTEETATEN]: z.string(),
     })
     .extend(commonEnvSchema.shape)