+++ apps/sif-demo-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @navikt/sif-demo-app
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [062d83e]
+- Updated dependencies [062d83e]
+    - @sif/soknad-ui@0.0.3
+    - @sif/api@0.0.2
+    - @navikt/k9-brukerdialog-prosessering-api@0.3.14
+    - @sif/soknad-forms@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes