+++ packages/sif-soknad-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sif/soknad-ui
 
+## 0.0.3
+
+### Patch Changes
+
+- 062d83e: Utvide med oppgavebekreftelse for aktivitetspenger + tilhørende refactoring
+
 ## 0.0.1
 
 ### Patch Changes