+++ packages/ung-innsyn/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @sif/ung-innsyn
 
+## 0.0.2
+
+### Patch Changes
+
+- 062d83e: Utvide med oppgavebekreftelse for aktivitetspenger + tilhørende refactoring
+- Updated dependencies [062d83e]
+- Updated dependencies [062d83e]
+    - @sif/api@0.0.2
+    - @navikt/k9-brukerdialog-prosessering-api@0.3.14
+    - @navikt/ung-brukerdialog-api@0.0.4
+
 ## 0.0.1
 
 ### Patch Changes